Adds Signup Email Process to Verify Account
Adds two new fields to the user model and generates their migrations.
These new fields are a verified boolean and a verify token. Two
lifecycle hooks are implemented in the user model, those are the
beforeCreate to generate the UUID token and the afterCreate to send
the account verify email. Additionally, adds a new route to
api/v1/users/verify/token so accounts can be verified. Adds a final
check during the login so that non verified accounts cannot log in.