Adds Features for Sending Forgot Password
Adds new field for a verify password token to the user model and creates
the appropriate migration. Creates a new /users/forgotPassword route
which post request will send an email with a verify password link.
TODO: Add a PATCH route which will require this token to update the
user's password.