Adds a Forgot Password Route
Adds a new route to users which with a valid token allows a user to
reset their password after it was forgotten. Forgotten password tokens
must first be obtained through a post to users/forgotPassword. Patch
requests to the same url with a username, password, password confirm
and token will result in changing the password.